Changes
Developed each of the four functions to work according to the instructions provided in the readme file.
A simple while loop menu was added for the user to interact with.

Edited the "name" variable to "country_name" within the .csv file as there was a conflict with its name when trying to code the functions.